+++ test/text_shaping/test_text_shaping.py
@@ -163,7 +163,7 @@ def test_text_with_parentheses(tmp_path):
     pdf.add_page()
     pdf.add_font(family="SBL_Hbrw", fname=HERE / "SBL_Hbrw.ttf")
     pdf.set_font("SBL_Hbrw", size=30)
-    pdf.set_text_shaping(30)
+    pdf.set_text_shaping(True)
     pdf.cell(text="אנגלית (באנגלית: English) ה", new_x="LEFT", new_y="NEXT")
     pdf.ln()
     pdf.cell(text="אנגלית (באנגלית: English) ", new_y="NEXT")